refactor: gnb data change to directory (#65)
