--- src/lib/contants.ts
@@ -1,30 +1,0 @@
-export const GNB_DATA = [
-  {
-    text: "스터디룸",
-    rootLink: "/study-room",
-    link: "/study-room/list",
-    icon: "/svg/ic-gnb-study-room.svg",
-    disabledIcon: "/svg/ic-gnb-study-room-disabled.svg",
-  },
-  {
-    text: "탐색",
-    rootLink: "/study/explorer",
-    link: "/study/explorer",
-    icon: "/svg/ic-gnb-search.svg",
-    disabledIcon: "/svg/ic-gnb-search-disabled.svg",
-  },
-  {
-    text: "커뮤니티",
-    rootLink: "#",
-    link: "#",
-    icon: "/svg/ic-gnb-community.svg",
-    disabledIcon: "/svg/ic-gnb-community-disabled.svg",
-  },
-  {
-    text: "마이",
-    rootLink: "/mypage",
-    link: "/mypage",
-    icon: "/svg/ic-gnb-my.svg",
-    disabledIcon: "/svg/ic-gnb-my-disabled.svg",
-  },
-];
